small improvements
- refractored everything
- added requirements.txt
- made whole codebase cleaner
- added classes
- made the overall file structore better to navigate
- removed duplicated functions
- changed function names to the normal convention
- moved files into the right folder
- added usage in readme.md
- added .gitignore
and more
