+++ .gitignore
@@ -1,0 +1,3 @@
+Result/*.mp4
+Videos/*.mp4
+__pycache__
