+++ README.md
@@ -1,2 +1,7 @@
 # JschlattReactionsMaker
 Makes Jschlatt React to your cursed videos
+
+# Usage
+1. Put your videos in the Videos folder
+2. run `pip install -r requirements.txt` (you only need to do this once)
+3. run `py main.py`
