--- Videos/PutYourClipsHere.txt
@@ -1,1 +1,0 @@
-
