+++ core/utils.py
@@ -1,0 +1,11 @@
+import os 
+
+def list_video_files(folder_path):
+    video_extensions = ['.mp4', '.avi', '.mkv', '.mov', '.flv', '.wmv', '.webm']
+    video_files = []
+
+    for file_name in os.listdir(folder_path):
+        if any(file_name.lower().endswith(ext) for ext in video_extensions):
+            video_files.append(file_name)
+
+    return video_files
