+++ main.py
@@ -1,0 +1,4 @@
+from core.gui import Gui
+
+if __name__ == "__main__":
+    Gui()()
